NEW Access to opauth config made transparent
At the moment, to change opauth config settings, we have to expose them
via the authenticator class and pass them through to Opauth.

This change allows the dev to directly set the opauth config, giving
them much better control and making it more transparent.

We're going to align better with Opauth's docs and convention if we do
this.