Fixing case where Member exists in DB already, but no session identity
OpauthController::finished() will check if the Member already exists,
write the identity, but not set OpauthIdentityID in session in this
case. Currently it only does that if the Member doesn't exist.

Because we're using the login in such a way that if even if the Member
exists, but isn't valid to log in, we want the OpauthIdentityID to be
set in the case of the Member existing as well to pre-populate forms
without the Member being logged in.

Note that OpauthController::loginAndRedirect() will eventually be called
after finished() and will clear OpauthIdentityID if it's able to log the
Member in successfully.